refactor: combine minimum virtual package and system requirements
